Use a newer version of Selenium
After several hours of testing, I updated the Selenium as far
as I dare. Nothing is broken.

The next version of Selenium is 4.0, and this appears to be the
last 3.x version (the version number converges towards pi).

This is a test commit. It can break something and if it does,
I will revert it.